Move EnemyClassConfig Class to Data Folder
