+++ Source/Aura/Public/Data/EnemyClassConfig.h
